use general confirmation dialog instead of just for DeleteConfirm
